Add rewind feature
It allows the player to rewind one turn.
The stack can grow up to 2048 turns. Once
that limit is reach it is cut in halves.

The game logic is unchanged but its code was
almost completly re-architectured. What a
mess of Some(Some((()))) it became. But on
the other hand I find it quite smart that the
step functions return the same "Update" objects
whether it represents the next step of the player
of the "opposite move" to store in the rewind
stack.