Fix snowballs not growing upon landing on snow
Not remembering the rules well, I programmed the snowballs to
keep the same size when being pushed off a bigger ball
onto a snowy tile.

It turns out the original game does not behave this way! The snow
ball grows if it lands on snow. No wonder I couldn't solve Adam.
(whitout exiting and re-entering the level from the other way.)