Add the reset feature to restart a level
The new feature is handy sometimes, but is not perfect.
It resets the current level, regardless of the state of
the others. That's something the rewind feature couldn't
achieve.

However the rewind queue does not store such an action.
So unlike the original game, to reset a level erases
completely the rewind queue. There is no coming back.
And the player's position is not reset, since the game
doesn't track it. So a snow ball might appear on the
player, or he might abuse the feature to get behind a
wall of snow balls and restart the level in a normally
impossible starting location.

Any level reset is recorded with a T in the save.txt
so that the game can still be saved and loaded.

Implementation-wise: a graph search is operated on the
initial map (stored in the process' static, read-only
memory) to get the current level's tiles. Then a diff
is generated between the initial and the current maps.